refactor: review feedback — prefer* param names, readable DAO batching, split test file
